Major refactoring to geometries and intersections
Overlaps collisions have become more specific, instead of being based
on the bounding volume of the other object.

In order to implement more intersection tests a Particle trait has
been added to represent point particles in motion. Additionally, types
that implement Intersects are now required to implement Particle, so
the operation is no longer commutative.

Intersects has been implemented for Compound and StaticBody but not
yet for RigidBody as Intersection<Moving<Capsule>> is not implemented
for Particle.

BVH no adds a new method, raycast, for performing such operations on
the BVH.

The Polygon trait has been completely refactored. Instead of
puzzlingly having a face method that returns a reference to
themselves, objects that implement Polygon are now simply required to
implement Contains<Point3<f32>> and Into<Plane>.  num_vertices and
num_edges have been removed and merged into the associated constant
NUM_VERTICES, which specifies both because the values are
equal. Hopefully in the future we can removed the edge method if
associated constants are allowed to be arrays of size specified by
another associated constant.