Huge number of changes, fixed physics, added example
I fixed a glaring error in physics resolution - along with changing
its structure to be more similar to typical constraint-based resolvers
- that has plagued the stability of my projects for a few months now.
It was hard to spot.

Added a demo that works fine under mgf_demo. Be sure to build it in
release mode. It's pretty fast, comparable to nphysics for sure, uses
entirely continuous collision detection, and reasonably stable.

I'll release 0.3, along with a demo video, on Monday.