Complete in-progress refactor of the physics system
The physics system has completely changed. I modified the constraint
solver to be more generic and allow for it to modifies array types
based on indices instead of mutable references.

In addition I added a RigidBodyVec class that handles physics in a
much more cache friendly way, and allows for a lot more tasks to
become open to multicore support. I am going to investigate that in
the future, but for now the most easily seen change is in mgf_demo,
which is much much faster.